ServerlessRAG_BedrockAgents/requirements.txt
# Core dependencies for local development and testing
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.34
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0

# AWS SDK
aioboto3==12.3.0

# Logging
python-json-logger==2.0.7
